Refactor to remove commissioner initialization from AppImpl to TvApp.

+++ examples/tv-app/android/java/AppPlatformShellCommands-JNI.cpp
@@ -48,7 +48,7 @@ static CHIP_ERROR pairApp(bool printHeader, size_t index)
     if (printHeader)
     {
         char str[64];
-        sprintf(str, "udc-commission %ld\r\n", static_cast<long>(index));
+        sprintf(str, "udc-commission %zu\r\n", index);
         strcat(response, str);
     }
 
@@ -57,7 +57,7 @@ static CHIP_ERROR pairApp(bool printHeader, size_t index)
     if (state == nullptr)
     {
         char str[64];
-        sprintf(str, "udc client[%ld] null \r\n", static_cast<long>(index));
+        sprintf(str, "udc client[%zu] null \r\n", index);
         strcat(response, str);
     }
     else
